Add timezone to env for java -Duser.timezone
Docker container environment defaults to UTC time. The image
files and logs are written out to a host filesystem with
America/New_York Timezone in current users systems.  Defaulting
ENV TZ to that for time being as all users are currently in that TZ.
This can be moved to a settings config file parameter as needed
in the future.